+++ src/main/java/com/geekster/employeeManagementSystem/repositories/IEmployeeDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IEmployeeDao extends JpaRepository<Employee,Long>  {
+public interface IEmployeeDao extends JpaRepository<Employee,Integer>  {
 
     Employee findFirstByEmployeeEmailId(String employeeEmailId);
 }
